requirements.txt
torch>=2.0.0
torchvision>=0.15.0
pillow>=8.0.0
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0
matplotlib>=3.5.0
tqdm>=4.60.0
joblib>=1.1.0
scipy>=1.7.0
pytorch-grad-cam>=1.5.0
